Access restrictions

+++ Controllers/Resources/CakePriceDTO.cs
@@ -7,7 +7,7 @@ namespace NextSugarCat.Controllers.Resources
 {
     public class CakePriceDTO
     {
-        public int CakePricePerPound { get; set; }
+        public double CakePricePerPound { get; set; }
         public decimal CakePricePerKg
         {
             get => Math.Round((decimal)(CakePricePerPound / 0.45359237), 2);
