+++ Controllers/Resources/ItemPricePerSetDTO.cs
@@ -8,7 +8,7 @@ namespace NextSugarCat.Controllers.Resources
     public class ItemPricePerSetDTO
     {
         public int? Id { get; set; }
-        public int SetPrice { get; set; }
+        public double SetPrice { get; set; }
         public int SetSize { get; set; }
     }
 }
