+++ Controllers/Resources/MenuItemSaveDTO.cs
@@ -8,7 +8,7 @@ namespace NextSugarCat.Controllers.Resources
         public string Name { get; set; }
         public string Type { get; set; }
         public string Description { get; set; }
-        public int? MinimumWeight { get; set; }
+        public double? MinimumWeight { get; set; }
         public MenuItemPriceDTO Price { get; set; }
         public ICollection<int> Ingredients { get; set; }
         public MenuItemSaveDTO()
